Restore the correct verify_integrity function
This function is the original one and the rodata verifying function
was moved to a separate function so that both behaviors can be
preserved. However during the rebase a mixup brought part of the
older changes back into the verify_integrity function.

This patch should be squashed into 0033-FIPS-embed-hmac.patch on a
rebase of the tree.

Signed-off-by: Simo Sorce <simo@redhat.com>